+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 - **[Change]** Match flag order for `DefineFont` and `DefineFontInfo`.
 
+## Rust
+
+- **[Breaking change]** Use bitflags instead of separate `bool` fields in `Bevel`, `ButtonCond`, `ButtonRecord`,
+  `ClipEventFlags`, `DefineDynamicText`, `DefineFont`, `DefineFontInfo`, `FileAttributes`, `Glow`, `GradientBevel`,
+  `GradientGlow`, `LineStyle` and `MorphLineStyle`.
+
 # 0.14.0 (2022-05-07)
 
 ## Rust